update looper and rules

+++ withdraw_looper.py
@@ -15,10 +15,14 @@ async def withdraw_ufo(user_id, address, count):
     if response is not None:
         print(response)
         print(f"💸 Вывод {count} UFO на кошелёк {address} прошел успешно | {user_id}\n{response}")
+        
+        return True
     else:
         print(response)
         print(f"😢 Ошибка отпраки токенов на кошелёк: {address} | {user_id}" + "\n" + 
                               "Сообщите в тех. поддержку")
+
+        return False
 
 async def main():
     while True:
@@ -28,8 +32,9 @@ async def main():
             await asyncio.sleep(30)
 
         for user in queue:
-            await withdraw_ufo(user[1], user[2], user[3])
-            await withdraw_db.delete_withdraw(user[1])
+            withdraw = await withdraw_ufo(user[1], user[2], user[3])
+            if withdraw:
+                await withdraw_db.delete_withdraw(user[1])
             await asyncio.sleep(5)
 
         await asyncio.sleep(30)
